Round unix timestamps for `received_ts`, `validfrom`, `validto` to nearest second and use python `int` type to match SQLite INTEGER type
- This is more for strict correctness than anything else as floats are sometimes casted to int type by SQLite if they end in .0 (hence can be perfectly represented as integers without precision loss).
- Additionally, sub-second precision is not necessary.

+++ esv_reference_server/msg_box/models.py
@@ -58,8 +58,8 @@ class MsgBoxAPITokenRow(NamedTuple):
     description: str
     canread: bool
     canwrite: bool
-    validfrom: float  # time.time()
-    validto: Optional[float]  # time.time()
+    validfrom: int  # time.time()
+    validto: Optional[int]  # time.time()
 
 
 class Message(NamedTuple):
@@ -67,7 +67,7 @@ class Message(NamedTuple):
     msg_box_api_token_id: int
     content_type: str
     payload: bytes
-    received_ts: float  # time.time()
+    received_ts: int  # time.time()
 
 
 class MessageMetadata(NamedTuple):
